array input into GUIs
